(update) Share Images for Release Notes
The images added in the previous commit have been updated to add a
`-twitter` to the file name. This is to ensure these images are used for
Twitter Cards and not the Open Graph images. The Open Graph images are
labeled with the `-og` suffix. The dimensions on these images differ
slightly to ensure they are optimized for the platform they are being
used on.